refactor: remove AlertMessageFormatter
